qa/cephfs: lua to respect missing kernel in yaml
When teuthology-suite is called with '-k none' option, which is valid,
there is no kernel record in job config created.

However at some test cases the lua premerge dies with exception:
  KeyError: 'kernel'
and when branch is not set for '-k none' and kernel client is
overridden:
  KeyError: 'branch'
so teuthology-suite quits unexpectedly without scheduling any jobs.

Fixes: https://tracker.ceph.com/issues/73676
Signed-off-by: Kyr Shatskyy <kyrylo.shatskyy@clyso.com>