qa/suites/upgrade: add "OBJECT_UNFOUND" to ignorelists
The thrashing in the upgrade tests has been configured to be very aggressive;
the tests are permitted to stop up to 4 of the 8 OSDs, so it is expected that
it is causing these kinds of health warnings to be generated.

This commit also cleans up some expected filesystem and pg peering warnings
in the upgrade tests.

Fixes: https://tracker.ceph.com/issues/72424
Signed-off-by: Laura Flores <lflores@ibm.com>